marketplace_builder_core::hooks
Trait BuilderHooks

pub trait BuilderHooks<Types: NodeType>:
    Sync
    + Send
    + 'static {
    // Provided methods
    fn process_transactions<'life0, 'async_trait>(
        &'life0 self,
        transactions: Vec<Types::Transaction>,
    ) -> Pin<Box<dyn Future<Output = Vec<Types::Transaction>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn handle_hotshot_event<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _event: &'life1 Event<Types>,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

A trait for hooks into the builder service. Used to further customize builder behaviour in ways not possible in builder core. If you don’t need such customisation, use NoHooks.

A simple example filtering incoming transactions based on some imaginary application-specific magic byte:

use marketplace_builder_core::hooks::BuilderHooks;
struct MyBuilderHooks { magic: u8 };

#[async_trait::async_trait]
impl BuilderHooks<MyTypes> for MyBuilderHooks {
    async fn process_transactions(&self, transactions: Vec<MyTransaction>) -> Vec<MyTransaction> {
        transactions.into_iter().filter(|tx| tx.bytes()[0] == self.magic).collect()
    }
}

Provided Methods

fn process_transactions<'life0, 'async_trait>( &'life0 self, transactions: Vec<Types::Transaction>, ) -> Pin<Box<dyn Future<Output = Vec<Types::Transaction>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Implement this to process transactions before they’ll be passed on to the builder.

fn handle_hotshot_event<'life0, 'life1, 'async_trait>( &'life0 self, _event: &'life1 Event<Types>, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Handle any hotshot event before the builder event loop handles it. Event handling is done sequentially, i.e. you can rely on the fact that the builder will process this event after the hooks have finished processing it. Accordingly, if this property is not important to you, and especially if you’re doing anything involved with this hook it is advisable to spawn a task doing the actual work and return, so that builder’s event loop isn’t blocked for too long.